refactor(tui): extract the funding declaration cluster into its own module
`declareFunding`, and the two collaborators it alone calls — `describe` and
`isAffirmative` — leave `import-orders.ts` for `import-orders-funding-declaration.ts`
(#223). Their only caller in that file was each other and the single call site in
`importBitgetOpenOrders`; verified by grep before the move, not assumed.

THE MOVE IS BYTE-IDENTICAL. The moved region diffs against its source by exactly the
`export` keyword on `declareFunding` — no re-wrapping, no renames, no string changes.
Docstrings moved whole, including the "once per BATCH, then overridable per order"
paragraph that carries the granularity argument.

`describe` and `isAffirmative` stay module-private: every rule they own is reachable
through `declareFunding`'s prompt channel, so neither needs to be named to be held.

THE HONEST DIFFERENCE FROM #211, #213, #219 AND #221: this one is not pure. It takes
the IO bag and awaits `io.ask`, so its test buys a scripted-prompt stub rather than a
plain call over a literal. What the extraction buys is the seam — the prompt strings
and the answer rules become drivable without an export file, a sidecar, a frozen clock
or a funding guard. The coverage argument is the sharpest in the series; the structural
one is the weakest. Both are stated in the new module's docstring.

`isAffirmative` remains duplicated in `record-fill.ts` deliberately. De-duplicating it
means choosing a shared home for a TUI-wide prompt primitive — a decision, not a move —
so neither module imports the other.

`import-orders.ts`: 637 → 592 lines. Suite unchanged at 1298 passed / 19 skipped,
typecheck exit 0, `module-graph.test.ts` passes unmodified.
